new train test, AND config for data split and log training
