+++ pipeline/configs/rdkit_biopy_config.yaml
@@ -6,13 +6,15 @@ proteins_data_path: "data/Protein-Sequence-Table.txt"
 glycan_encoder_type: "rdkit"
 protein_encoder_type: "biopy"
 binding_predictor_type: "dummy"
-k_folds: 8
-num_epochs: 5
+k_folds: 3
+num_epochs: 3
 batch_size: 32
-learning_rate: 0.001
+learning_rate: 0.1
 checkpoint_frequency: 2
 train_final_model: False # if want to train and save the model on the best found params from cross val usinmg the entire training set.
+log_predict: True # tranform our target to log(1+target) (log1p()) so that we dont train the model to predict all zeros (reverted on predictions output with expm1())
 random_state: 42
+split_mode: 'AND'
 # optional
 val_split: 0.2
 device: "cpu"
