+++ pipeline/src/utils/config.py
@@ -14,14 +14,15 @@ class TrainingConfig:
     glycan_encoder_type: str
     protein_encoder_type: str
     binding_predictor_type: str
+    split_mode: str
     k_folds: int
     num_epochs: int
     batch_size: int
     learning_rate: float
     checkpoint_frequency: int
     random_state: int
+    log_predict: bool
     train_final_model: bool
-    
     # optional parameters
     val_split: float = 0.2
     device: str = "cuda"  # or "cpu"
